fixed duplicate hp bug and added random amounts of random options

+++ src/main/java/data/engines/Engine.java
@@ -2,11 +2,11 @@ package data.engines;
 
 public abstract class Engine {
 
-    private int engineCode;
+    private long engineCode;
     private int hp;
     private int torque;
 
-    public Engine(int engineCode, int hp, int torque) {
+    public Engine(long engineCode, int hp, int torque) {
         this.engineCode = engineCode;
         this.hp = hp;
         this.torque = torque;
